+++ packages/remix-dev/compiler/loaders.ts
@@ -22,6 +22,7 @@ export const loaders: { [ext: string]: esbuild.Loader } = {
   ".ogg": "file",
   ".otf": "file",
   ".png": "file",
+  ".sql": "text",
   ".svg": "file",
   ".ts": "ts",
   ".tsx": "tsx",