+++ packages/remix-dev/config.ts
@@ -1,4 +1,5 @@
 import * as path from "path";
+import { pathToFileURL } from "url";
 import * as fse from "fs-extra";
 import getPort from "get-port";
 
@@ -287,15 +288,28 @@ export async function readConfig(
   }
 
   let rootDirectory = path.resolve(remixRoot);
-  let configFile = path.resolve(rootDirectory, "remix.config.js");
+  let configFile = findConfig(rootDirectory, "remix.config");
 
-  let appConfig: AppConfig;
-  try {
-    appConfig = require(configFile);
-  } catch (error) {
-    throw new Error(
-      `Error loading Remix config in ${configFile}\n${String(error)}`
-    );
+  let appConfig: AppConfig = {};
+  if (configFile) {
+    let appConfigModule: any;
+    try {
+      // shout out to next
+      // https://github.com/vercel/next.js/blob/b15a976e11bf1dc867c241a4c1734757427d609c/packages/next/server/config.ts#L748-L765
+      if (process.env.NODE_ENV === "test") {
+        // dynamic import does not currently work inside of vm which
+        // jest relies on so we fall back to require for this case
+        // https://github.com/nodejs/node/issues/35889
+        appConfigModule = require(configFile);
+      } else {
+        appConfigModule = await import(pathToFileURL(configFile).href);
+      }
+      appConfig = appConfigModule?.default || appConfigModule;
+    } catch (error) {
+      throw new Error(
+        `Error loading Remix config at ${configFile}\n${String(error)}`
+      );
+    }
   }
 
   let customServerEntryPoint = appConfig.server;
@@ -471,6 +485,17 @@ function findEntry(dir: string, basename: string): string | undefined {
   for (let ext of entryExts) {
     let file = path.resolve(dir, basename + ext);
     if (fse.existsSync(file)) return path.relative(dir, file);
+  }
+
+  return undefined;
+}
+
+const configExts = [".js", ".cjs", ".mjs"];
+
+function findConfig(dir: string, basename: string): string | undefined {
+  for (let ext of configExts) {
+    let file = path.resolve(dir, basename + ext);
+    if (fse.existsSync(file)) return file;
   }
 
   return undefined;